audiototext/requirements.txt
# Основные зависимости для преобразования аудио в текст
openai-whisper==20231117
flask==2.3.3
werkzeug==2.3.7

# Дополнительные зависимости для обработки аудио
torch>=1.9.0
torchaudio>=0.9.0
numpy>=1.21.0

# Для работы с различными форматами аудио
ffmpeg-python==0.2.0
pydub==0.25.1

# Для веб-интерфейса
jinja2==3.1.2
